pyproject.toml
[project]
name = "olx-scraper"
version = "0.1.0"
description = "olx scraper"
authors = [{ name = "Aleksander Drwal", email = "olek.drwal@gmail.com" }]
dependencies = [
    "gql>=3.5.3",
    "pydantic>=2.11.5",
    "typer>=0.16.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "returns[compatible-mypy]>=0.25.0",
    "mypy>=1.15.0",
    "requests-toolbelt>=1.0.0",
    "types-requests>=2.32.0.20250602",
    "types-psycopg2>=2.9.21.20250516",
    "numpy>=2.3.0",
]
requires-python = ">=3.13"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.ariadne-codegen]
# schema_path = "schema.graphql"     # Or an introspection URL
queries_path = "schema.graphql" # Or a list of query files

[tool.mypy]
plugins = ["returns.contrib.mypy.returns_plugin"]
